+++ NEWS.md
@@ -3,7 +3,7 @@
 * "getting started" vignette instead of the long README.
 * get back functions reference table in the README.
 * expose the `dir_size` function.
-* add new functions `shiny_app_size` and `shiny_app_deps`.
+* add new functions `app_size` and `app_deps`.
 * new Imports dependency `renv`.
 
 # pacs 0.3.8